+++ packages/excalidraw/components/LayerUI.scss
@@ -1,6 +1,6 @@
 @import "open-color/open-color";
 @import "../css/variables.module.scss";
-@import './EmojiReactions.scss';
+@import './emojiReactions/EmojiReactions.scss';
 
 .excalidraw {
   .layer-ui__wrapper.animate {
