+++ packages/excalidraw/components/emojiReactions/EmojiReactions.scss
