+++ packages/excalidraw/components/emojiReactions/FloatingEmoji.tsx
