+++ packages/excalidraw/components/emojiReactions/index.ts
@@ -1,0 +1,12 @@
+export { useEmojiReactions } from "./useEmojiReactions";
+export { ReactionOverlay } from "./ReactionOverlay";
+export { FloatingEmojisLayer } from "./FloatingEmojisLayer";
+export { FloatingEmoji } from "./FloatingEmoji";
+export { EmojiPickerPanel } from "./EmojiPickerPanel";
+export { ReactionModeButton } from "./ReactionModeButton";
+export { default as ReactionEmojiSubmenu } from "./ReactionEmojiSubmenu";
+
+export type {
+  FloatingEmojiData,
+  UseEmojiReactionsResult,
+} from "./useEmojiReactions";
